+++ inst/runit_tests/expected_files.r
@@ -1,16 +1,17 @@
-require("document")
-test_txt <- function() {
+if (interactive()) pkgload::load_all()
+test_html <- function() {
     options(useFancyQuotes = FALSE)
-    output_directory <- file.path(tempdir(), "document_runit_txt")
+    output_directory <- file.path(tempdir(), "document_runit_html")
     unlink(output_directory, recursive = TRUE)
     dir.create(output_directory)
-    file_name  <- file.path(system.file("expected_files",
-                                        package = "document"), "simple.txt")
-    input <- file.path(system.file("files", package = "document"),
-                       sub("\\.txt", ".R", basename(file_name)))
-    txt_name <- document(input, output_directory = output_directory,
-                         check_package = FALSE, runit = TRUE)[["txt_path"]]
-    current <- readLines(txt_name)
-    reference  <- readLines(file_name)
+    input <- file.path(system.file("files", package = "document"), "simple.R")
+    output <- document(input, output_directory = output_directory,
+                       check_package = FALSE, runit = TRUE)
+    # html
+    ref_path <- file.path(system.file("expected_files", package = "document"),
+                          sub("\\.R", ".html", basename(input)))
+    reference  <- readLines(ref_path)
+    output_path <- output[["html_path"]]
+    current <- readLines(output_path)
     RUnit::checkTrue(identical(current, reference))
 }