+++ tests/testthat.R
@@ -1,5 +1,5 @@
-document::set_RUN_R_TESTS(document::is_running_on_fvafrcu_machines() ||
-                          document::is_running_on_gitlab_com(), force = TRUE)
+fritools::set_run_r_tests(fritools::is_running_on_fvafrcu_machines() ||
+                          fritools::is_running_on_gitlab_com(), force = TRUE)
 library("testthat")
 library("document")
 test_check("document")